[IMP] Added more improvements about QA for Python and Javascript source code to split each class to separate file, that helps with logic analysis and navigation, also rename some classes, fuctions and variables names more standard.
