+++ __manifest__.py
@@ -22,18 +22,18 @@ Partner And Google Map Snippet compatible.
         'website',
     ],
     'data': [
-        'data/yandex_maps_data.xml',
         'data/website_shops_map_data.xml',
+        'data/yandex_maps_data.xml',
         'security/ir.model.access.csv',
-        'views/shop_property_view.xml',
-        'views/extend_module_view.xml',
-        'views/shop_api_user_view.xml',
-        'views/shop_map_cache_view.xml',
-        'views/period_of_time_view.xml',
+        'views/ir_module_module_view.xml',
         'views/res_config_settings.xml',
-        'views/stock_view.xml',
-        'views/shop_list_config_view.xml',
         'views/s_shop_list.xml',
+        'views/stock_view.xml',
+        'views/shop_api_user_view.xml',
+        'views/shop_list_config_view.xml',
+        'views/shop_map_cache_view.xml',
+        'views/shop_property_view.xml',
+        'views/shop_time_period_view.xml',
         'views/shop_work_days_view.xml',
         'views/website_frontend_assets.xml',
         'views/website_frontend_templates.xml',
