+++ models/__init__.py
@@ -1,9 +1,13 @@
-from . import eyekraft_shop
+from . import public_shop
 from . import inherit_stock
+from . import ir_module_module
 from . import res_config_settings
 from . import res_partner
 from . import shop_api_user
 from . import shop_list_config
 from . import shop_map_cache
 from . import shop_property
+from . import shop_time_period
+from . import shop_work_days
 from . import shop_work_hours
+from . import website
