+++ static/src/js/website.editor.js
@@ -18,11 +18,11 @@ odoo.define('website_shops_map.editor', function (require) {
             editor.editor_bar.prependTo(document.body);
             var messageList = _t("Select property");
 
-            var selector = $(document).find('#eyekraft_props_picker_selector')
+            var selector = $(document).find('#shop_props_picker_selector')
             if (selector.length){
                 // remove properties
                 $(selector).children().remove();
-                $(selector).append("<select id='eyekraft_props_picker' class='hidden' multiple='' title='+ messageList +'></select>");
+                $(selector).append("<select id='shop_props_picker' class='hidden' multiple='' title='+ messageList +'></select>");
             };
 
             var bigMapView = $(document).find('#map-container-vertical')
