+++ views/ir_module_module_view.xml
