+++ views/shop_time_period_view.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <!-- "Period of time" form view -->
-    <record id="period_of_time_view_form" model="ir.ui.view">
-        <field name="name">period.of.time.form</field>
-        <field name="model">period.of.time</field>
+    <!-- "Shop Time Period" form view -->
+    <record id="shop_time_period_view_form" model="ir.ui.view">
+        <field name="name">shop.time.period.form</field>
+        <field name="model">shop.time.period</field>
         <field name="arch" type="xml">
-            <form string="Period of time">
+            <form string="Shop Time Period">
                 <sheet>
                     <group>
                         <group>
@@ -20,17 +20,17 @@
         </field>
     </record>
 
-    <!-- "Period of time" ir.actions.act_window -->
-    <record id="period_of_time_action" model="ir.actions.act_window">
-        <field name="name">Periods of Time</field>
-        <field name="res_model">period.of.time</field>
+    <!-- "Shop Time Period" ir.actions.act_window -->
+    <record id="shop_time_period_action" model="ir.actions.act_window">
+        <field name="name">Shop Time Period</field>
+        <field name="res_model">shop.time.period</field>
         <field name="view_mode">tree,form</field>
         <field name="help" type="html">
             <p class="o_view_nocontent_smiling_face">
-                Create a Periods of Time.
+                Create a Shop Time Period.
             </p>
             <p style="text-align: left">
-                Create a Periods of Time entering opening hours for the shop
+                Create a Shop Time Period entering opening hours for the shop
                 <ul style="text-align: left">
                     <li><i>Since</i>, defines the hour from in which the shop works, in set hours in numbers</li>
                     <li><i>To</i>, defines the hour until in which the shop works, in set hours in numbers</li>
@@ -39,10 +39,10 @@
         </field>
     </record>
 
-    <!-- "Period of time" menu item -->
-    <menuitem id="period_of_time_menu"
-        action="period_of_time_action"
+    <!-- "Shop Time Period" menu item -->
+    <menuitem id="shop_time_period_menu"
+        action="shop_time_period_action"
         parent="menu_shop_config"
-        name="Periods of Time"
+        name="Shop Time Period"
         sequence="120" />
 </odoo>
