+++ views/shop_work_days_view.xml
@@ -2,8 +2,8 @@
 <odoo>
     <!-- "Shop Work Days" form view -->
     <record id="shop_work_days_view_form" model="ir.ui.view">
-        <field name="name">eyecraft.work.days.form</field>
-        <field name="model">eyecraft.work.days</field>
+        <field name="name">shop.work.days.form</field>
+        <field name="model">shop.work.days</field>
         <field name="arch" type="xml">
             <form string="Shop Work Days">
                 <sheet>
@@ -17,7 +17,7 @@
     <!-- "Shop Work Days" ir.actions.act_window -->
     <record id="shop_work_days_action" model="ir.actions.act_window">
         <field name="name">Shop Work Days</field>
-        <field name="res_model">eyecraft.work.days</field>
+        <field name="res_model">shop.work.days</field>
         <field name="view_mode">tree,form</field>
         <field name="help" type="html">
             <p class="o_view_nocontent_smiling_face">
